Generate link text for subsequent files with the same name
When a file with a name which matches an existing file in the file area is
uploaded, it is assumed to be a different (additional) file rather than
the same file.  The "newfile" information contains the filename in a
"filename" attribute rather than "file" (as used by the non-new file), so
links generated for these files had no text and were thus not visible.